requirements.txt
# インストールするライブラリを記載するテキストファイル
Flask==1.1.1
line-bot-sdk==1.15.0
beautifulsoup4==4.8.2
soupsieve==1.9.5
urllib3==1.25.8
opencv-python==4.2.0.32
xmltodict==0.12.0
selenium==3.141.
SpeechRecognition==3.8.1
pydub==0.23.1
psycopg2-binary==2.8.5
requests==2.22.0
pytz==2020.1
